actualized Index Buffer

+++ src/tools/QTDemoApp/RenderWidget.h
@@ -32,7 +32,7 @@ const std::vector<Vertex> vertices = {
 	{ { -0.5f, 0.5f },{ 1.0f, 1.0f, 1.0f } }
 };
 
-const std::vector<uint16_t> indices = {
+const std::vector<uint32_t> indices = {
 	0, 1, 2, 2, 3, 0
 };
